% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAClasses.R
\docType{class}
\name{featuresFrame-class}
\alias{featuresFrame-class}
\title{featureFrame class}
\description{
An S4 class contains the information of features used to fit a model
}
\section{Slots}{

\describe{
\item{\code{var.names}}{A character vector containing the name of variabes from the dataset used to generate the features}

\item{\code{feature.types}}{A list containing the definition of features}
}}

