\name{gui}

\alias{gui}
\alias{gui,ANY-method}

  
\title{Graphical User Interface}

\description{
Open a ShinyApp that is used to provide graphical user interface for fitting and evaluating species distribution models.
}

\usage{
\S4method{gui}{ANY}(...)

}

\arguments{
  \item{...}{not implemented yet.}
}
\details{
To run the function, several packages including shiny, shinythemes, shinyBS, and leaflet should be installed.
}

\value{
A HTML page in browser is opened.
}


\author{Babak Naimi \email{naimi.b@gmail.com}

\url{http://r-gis.net}

\url{http://biogeoinformatics.org}

}


\examples{
# gui()
}

\keyword{spatial}
