% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_version.R
\name{maxent_version}
\alias{maxent_version}
\title{Get Maxent version}
\usage{
maxent_version()
}
\value{
The version number of the installed Maxent program.
}
\description{
Get the version number of the installed Maxent program
}
\references{
This function is adapted from \code{dismo:::.getMeVersion}.
}
