\name{names}

\alias{labels,speciesData-method}
\alias{names}
\alias{names<-}
\alias{names,speciesData-method}
\alias{names,singleSpecies-method}
\alias{names,multipleSpecies-method}
\alias{names,SpeciesDataList-method}
\alias{names<-,speciesData,character-method}
\alias{names<-,speciesData-method}

  
\title{Names of species}

\description{
Get or set the names of the species of a speciesData object 
}

\usage{
\S4method{names}{speciesData}(x)

\S4method{names}{speciesData}(x)<-value

}

\arguments{
  \item{x}{speciesData (either of singleSpecies, multiple Species or SpeciesDataList) object}
  \item{value}{character (vector)}
}

\value{
Character 
}


\author{Babak Naimi \email{naimi.b@gmail.com}

\url{http://r-gis.net}

\url{http://biogeoinformatics.org}

}


\examples{
#
}

\keyword{spatial}
