% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAClasses.R
\docType{class}
\name{sdmFormula-class}
\alias{sdmFormula-class}
\title{sdmFormula Class}
\description{
An S4 class to represent different terms and component in formula
}
\section{Slots}{

\describe{
\item{\code{formula}}{The formula defined by user}

\item{\code{vars}}{The name of variables defined in the formula}

\item{\code{species}}{The name of species (response) variables}

\item{\code{model.terms}}{The defined terms in the formula that will be used in model fitting}

\item{\code{data.terms}}{The defined terms in the formula that will be used in data manipulation}
}}

