\name{evaluates}

\alias{evaluates}
\alias{evaluates,vector,vector-method}

\title{evaluate for accuracy}

\description{
evaluates for accuracy 
}

\usage{
evaluates(o,p,...)
}

\arguments{
  \item{o}{a numeric vector including the observed values}
  \item{p}{a numeric vector including the predicted values}
  \item{...}{additional arguments (not implemented)}
  }
\details{
evaluates the accuracy
}

\value{
 an object of class \code{sdmEvaluate}
}

\references{
Naimi, B., Araújo, M.B. (2016) sdm: a reproducible and extensible R platform for species distribution modelling, Ecography, DOI: 10.1111/ecog.01881
}

\author{Babak Naimi \email{naimi.b@gmail.com}

\url{http://biogeoinformatics.org}
}

\seealso{# }

\examples{
#
}

\keyword{spatial,species,data}